Code is a mess and needs uncommented but it mostly works
Backup of some work (aka shitty code that does it's job just like nenad)

+++ rocksat/MainCtrl/rocksat/imu.h
@@ -3,6 +3,6 @@
 
 
 void triggerIMUDataCollection();
-void getImuData(short* );
+void getImuData(byte* );
 
 #endif
